pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "dotenvironment"
version = "1.1.0"
authors = [
    {name = "Pável Hernández Reza", email = "phr2807@gmail.com" }]
description = "Pequeña librería diseñada para facilitar la obtención y organización de variables de entorno, casteos y valores predeterminados centralizando todo el flujo de trabajo en un solo lugar."
readme = "README.md"
license = "MIT"
dependencies = [
    "python-dotenv==1.2.1",
]

[tool.setuptools]
package-dir = { "" = "src" }
